--- web/lib/api/admin.ts
@@ -1,14 +1,0 @@
-import * as admin from 'firebase-admin'
-
-export function initAdmin() {
-  const serviceAccount = JSON.parse(
-    process.env.FIREBASE_SERVICE_ACCOUNT_KEY ?? ''
-  )
-  if (admin.apps.length === 0) {
-    admin.initializeApp({
-      credential: admin.credential.cert(serviceAccount),
-    })
-  }
-
-  return admin
-}
